feature(typography): Add Variable font and styles
I picked out a font from google fonts, but it's licensed to be used as a
webfont. For better performance, I want to host the fonts with the site
directly so they're not pulling from google on every page load.

Also add some basic styling with support for scaleable type.